requirements.txt
crawl4ai>=0.2.0
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
python-slugify>=8.0.1
typer>=0.9.0
pytest>=7.3.1
pytest-cov>=4.1.0
